+++ metadrive/envs/scenario_env.py
@@ -1,0 +1,3 @@
+"""
+This environment can load all scenarios exported from other environments via env.export_scenarios()
+"""
